VMS-52481: Fix free GPU memory check
Closes VMS-52481

(cherry picked from commit 4cd6444cfd50d5e24df86391a800187a65ee5804)

+++ artifacts/nvidia/Video_Codec_SDK_11.1.5/Samples/Utils/Resize.cu
@@ -205,7 +205,7 @@ void ScaleYUV420(unsigned char *dpDstY,
     }
 }
 
-int freeGpuMemory()
+size_t freeGpuMemory()
 {
     size_t free, total;
     cudaMemGetInfo(&free, &total);
